requirements.txt
selenium>=4.15.0
beautifulsoup4>=4.12.0
pyodbc>=5.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
pandas>=2.1.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
tenacity>=8.2.0
pytest>=7.4.0
httpx>=0.26.0
jinja2>=3.1.0
python-multipart>=0.0.6
schedule>=1.2.0
